Fixed init script.

+++ web/directory/management/commands/docker_init_db_data.py
@@ -8,12 +8,18 @@ class Command(BaseCommand):
 
     def handle(self, *args, **options):
         self.stdout.write('Deleting existing data')
-        CoopType.objects.all().delete()
-        Coop.objects.all().delete()
-        Address.objects.all().delete()
-        Locality.objects.all().delete()
-        State.objects.all().delete()
-        Country.objects.all().delete()
+        if CoopType.objects.exists():
+            CoopType.objects.all().delete()
+        if Coop.objects.exists():
+            Coop.objects.all().delete()
+        if Address.objects.exists():
+            Address.objects.all().delete()
+        if Locality.objects.exists():
+            Locality.objects.all().delete()
+        if State.objects.exists():
+            State.objects.all().delete()
+        if Country.objects.exists():
+            Country.objects.all().delete()
         self.stdout.write('Seeding initial data')
         call_command('loaddata', 'directory/fixtures/country_data.yaml')
         call_command('loaddata', 'directory/fixtures/state_data.yaml')
